+++ app/DataObjects/Cocktail/Ingredient.php
@@ -6,8 +6,11 @@ namespace Kami\Cocktail\DataObjects\Cocktail;
 
 class Ingredient
 {
+    /**
+     * @param array<int> $substitutes
+     */
     public function __construct(
-        public readonly ?int $id,
+        public readonly int $id,
         public readonly ?string $name,
         public readonly float $amount,
         public readonly string $units,