+++ routes/api.php
@@ -87,7 +87,6 @@ Route::middleware($authMiddleware)->group(function() {
 
     Route::prefix('cocktails')->group(function() {
         Route::get('/', [CocktailController::class, 'index'])->name('cocktails.index');
-        Route::get('/random', [CocktailController::class, 'random'])->name('cocktails.random');
         Route::get('/user-shelf', [CocktailController::class, 'userShelf'])->name('cocktails.user-shelf');
         Route::get('/{id}', [CocktailController::class, 'show'])->name('cocktails.show');
         Route::post('/{id}/toggle-favorite', [CocktailController::class, 'toggleFavorite'])->name('cocktails.favorite');