+++ CHANGELOG.md
@@ -12,6 +12,8 @@
     - **Upgrade guide**: Available via `/cocktails?filter[on_shelf]=true`
 - Removed `/ingredients/find` endpoint
     - **Upgrade guide**: Available via `/ingredients?filter[name]=whiskey`
+- Removed `/cocktails/random` endpoint
+    - **Upgrade guide**: n/a
 - Removed `site_search_index` indexing
     - **Upgrade guide**: If possible migrate to federated/multi-index search
 - Redis is now mandatory dependency